fix: the most recent PRs first to mitigate timeouts
The script engine will kill any scheduled job if it takes longer than 5min
to complete. Depending on the network load and number of PRs to be checked
it is possible that the cla script will be killed beefore it checks all of
the PRs.

Channging the order will at least allow us to check the most recent PRs first
and only if there is time left, we check the older PRs.

Caveat: since check and update all PRs in stages, this change might not be
sufficient, because even if we check the PRs the won't be labeled until
we know the state of all PRs.